Add call_nodes/3 and call_nodes_minimal/3 funs.
call_nodes/3 fun call all given nodes and executes the given request.
call_nodes_minimal/3 fun calls the minimum set of nodes to cover all shards and execute the given request. It aims to call one shard per node but it is not always possible. For example on a 3 node cluster with replication factor 2, we need to call at least 2 nodes to cover all shards, but half of the shards will be covered twice on by two calls.